temp_repo/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "discord-py>=2.5.2",
    "matplotlib>=3.10.1",
    "nltk>=3.9.1",
    "numpy>=2.2.4",
    "openai>=1.70.0",
    "pandas>=2.2.3",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.0",
    "pytz>=2025.2",
    "scipy>=1.15.2",
    "streamlit>=1.44.0",
    "trafilatura>=2.0.0",
    "twilio>=9.5.1",
    "yfinance>=0.2.55",
]
